Convert Product On Sale block file from JS to TS
